Split build script into separate modules
